Update policies individually
Try, except does not kick us out of statements upon first policy
failure, but rather tries each one individually.  This could happen if
a policy has not seen this article offered before (because it is brand
new).  When this happens, just skip it until it becomes part of the
policy’s possible actions a short time later.